refactor(core): move streamer extensions to streamer package
